No more shooting railgun when dead
note that it doesn't use the existing disable / on enable. Because it would (very likely) cause desyncrhonization issues. Fixing them require smth like a few hours of locking in and debugging and rewritting a big part of the architecture which I'm too lazy to do right sooo we'll stick with this version that works